requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohttp==3.8.6
    # via openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
cachetools==5.3.2
    # via google-auth
certifi==2023.7.22
    # via requests
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
colorama==0.4.6
    # via
    #   halo
    #   log-symbols
coverage==7.3.2
    # via pytest-cov
defusedxml==0.7.1
    # via jira
diskcache==5.6.1
    # via -r requirements.in
fire==0.5.0
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.12.0
    # via google-api-python-client
google-api-python-client==2.106.0
    # via -r requirements.in
google-auth==2.23.4
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.1
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.1.0
    # via -r requirements.in
googleapis-common-protos==1.61.0
    # via google-api-core
greenlet==3.0.0
    # via playwright
halo==0.0.31
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jira==3.8.0
    # via -r requirements.in
log-symbols==0.0.14
    # via halo
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
openai==0.27.8
    # via -r requirements.in
packaging==23.2
    # via
    #   jira
    #   pytest
pillow==10.3.0
    # via jira
playwright==1.39.0
    # via pytest-playwright
pluggy==1.3.0
    # via pytest
protobuf==4.25.0
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyee==11.0.1
    # via playwright
pyparsing==3.1.1
    # via httplib2
pytest==7.3.2
    # via
    #   -r requirements.in
    #   pytest-base-url
    #   pytest-cov
    #   pytest-playwright
pytest-base-url==2.0.0
    # via pytest-playwright
pytest-cov==4.1.0
    # via -r requirements.in
pytest-playwright==0.3.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python-slugify==8.0.1
    # via pytest-playwright
pytz==2024.1
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   jira
    #   openai
    #   pytest-base-url
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   jira
requests-toolbelt==1.0.0
    # via jira
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   fire
    #   halo
spinners==0.0.24
    # via halo
termcolor==2.3.0
    # via
    #   fire
    #   halo
text-unidecode==1.3
    # via python-slugify
tqdm==4.65.0
    # via
    #   -r requirements.in
    #   openai
typing-extensions==4.8.0
    # via
    #   jira
    #   pyee
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.0.7
    # via requests
yarl==1.9.2
    # via aiohttp
